Moves OLED code to varys.c, removes RGB code entirely, moves defines over to keyboard.json, removes default rules, adds features to keyboard.json
