+++ keyboards/varys/config.h
@@ -2,28 +2,4 @@
 // SPDX-License-Identifier: GPL-2.0-or-later
 
 #pragma once
-#define SPLIT_HAND_PIN D7
 #define SPLIT_HAND_PIN_LOW_IS_LEFT
-#define USE_I2C 
-#define SPLIT_USB_DETECT
-
-//#define SOFT_SERIAL_PIN D2
-
-#define ENCODERS_PAD_A { F6 }
-#define ENCODERS_PAD_B { F7 }
-
-#define SPLIT_OLED_ENABLE
-
-// #define IS31FL3737_I2C_ADDRESS_1 IS31FL3737_I2C_ADDRESS_GND
-// #define DRIVER_1_LED_TOTAL 28
-// #define DRIVER_2_LED_TOTAL 28
-// #define RGB_MATRIX_LED_COUNT (DRIVER_1_LED_TOTAL + DRIVER_2_LED_TOTAL)
-// #define RGB_MATRIX_SPLIT { DRIVER_1_LED_TOTAL, DRIVER_2_LED_TOTAL } 
-// #define RGB_MATRIX_MAXIMUM_BRIGHTNESS 120  // Lower maximum brightness (range is 0-255)
-// #define RGB_MATRIX_DISABLE_KEYCODES
-// #define USB_MAX_POWER_CONSUMPTION 500
-// #define RGB_MATRIX_STARTUP_VAL 120
-// #define RGBLIGHT_SPLIT
-// #define RGBLED_SPLIT { DRIVER_1_LED_TOTAL, DRIVER_2_LED_TOTAL }
-// #define SPLIT_TRANSPORT_MIRROR
-// #define I2C1_CLOCK_SPEED 400000
