--- keyboards/varys/rules.mk
@@ -1,5 +1,0 @@
-ENCODER_ENABLE = yes
-OLED_ENABLE = yes
-OLED_DRIVER = ssd1306
-OLED_TRANSPORT = i2c
-SPLIT_KEYBOARD = yes
