+++ keyboards/varys/varys.c
@@ -1,0 +1,64 @@
+// Copyright 2023 Vinay Komaravolu
+// SPDX-License-Identifier: GPL-2.0-or-later
+#include "quantum.h"
+
+enum layer_number {
+    _QWERTY = 0,
+    _RAISE,
+    _LOWER,
+    _MISC,
+};
+
+#ifdef OLED_ENABLE
+static const char PROGMEM varys_logo[] = {
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xc0, 0xe0, 0xf0, 0xf0, 0xf8, 
+0x38, 0x18, 0x38, 0x78, 0xf8, 0x70, 0x70, 0xe0, 0xc0, 0xc0, 0x80, 0x80, 0x80, 0x00, 0x00, 0x00, 
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xe0, 0xfc, 0xff, 0xff, 0xef, 0x1f, 0x3f, 0x7f, 
+0xff, 0xfe, 0xfe, 0xff, 0xc6, 0x02, 0x03, 0x01, 0x01, 0x00, 0x00, 0x01, 0x01, 0x01, 0x00, 0x00, 
+0x00, 0x00, 0x00, 0x00, 0x80, 0xf0, 0xfc, 0xff, 0x0f, 0xf7, 0xfb, 0xfd, 0x66, 0x37, 0xdf, 0xee, 
+0xe4, 0x71, 0xb3, 0xc7, 0xff, 0x7f, 0xfe, 0xf8, 0xe0, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+0x00, 0x00, 0x00, 0xf0, 0xff, 0xff, 0xc1, 0xfe, 0xff, 0x03, 0xff, 0xff, 0x00, 0xff, 0xff, 0x01, 
+0xfe, 0xff, 0xc3, 0x99, 0x3c, 0x76, 0xef, 0x9f, 0xbf, 0xff, 0xf8, 0x00, 0x00, 0x00, 0x00, 0x00, 
+0x00, 0x00, 0x00, 0x7f, 0x0f, 0x07, 0xff, 0xff, 0xff, 0x00, 0xff, 0xff, 0x00, 0xff, 0xff, 0x80, 
+0x7f, 0xff, 0xfd, 0xfb, 0x03, 0xfc, 0xff, 0xff, 0xff, 0x0f, 0x7f, 0x7e, 0x00, 0x00, 0x00, 0x00, 
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x07, 0x1f, 0x7f, 0xfc, 0xf9, 0xf3, 0xe7, 0xcc, 0x99, 0x73, 
+0x67, 0xee, 0xed, 0xf1, 0xfc, 0xff, 0x3f, 0x0f, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x06, 0x06, 0x03, 0x07, 0x07, 0x04, 0x00, 0x00, 0x00, 
+0x00, 0x00, 0x06, 0x07, 0x07, 0x06, 0x06, 0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+  };
+
+oled_rotation_t oled_init_user(oled_rotation_t rotation) {
+    return OLED_ROTATION_270;  // flips the display 180 degrees if offhand
+}
+
+bool oled_task_kb(void) {
+    oled_set_cursor(0, 1);
+    // Draw Logo 32x96
+    oled_write_raw_P(varys_logo, sizeof(varys_logo));
+    oled_set_cursor(0, 14);
+    switch (get_highest_layer(layer_state)) {
+      case _QWERTY:
+        oled_write_P(PSTR("VARYS"), false);
+        break;
+      case _LOWER:
+        oled_write_P(PSTR("LOWER"), false);
+        break;
+      case _RAISE:
+        oled_write_P(PSTR("RAISE"), false);
+        break;
+      case _MISC:
+        oled_write_P(PSTR("MISC "), false);
+        break;
+      default:
+        oled_write_ln_P(PSTR("Undefined"), false);
+    }
+    led_t led_state = host_keyboard_led_state();
+    oled_set_cursor(0, 9);
+    oled_write_P(led_state.num_lock ? PSTR(" NUM ") : PSTR("     "), false);
+    oled_set_cursor(0, 10);
+    oled_write_P(led_state.caps_lock ? PSTR(" CAP ") : PSTR("     "), false);
+    oled_set_cursor(0, 11);
+    oled_write_P(led_state.scroll_lock ? PSTR(" SCR ") : PSTR("     "), false);
+    return false;
+}
+#endif
